requirements.txt
customtkinter==5.2.2
pvporcupine==4.0.1
pvrecorder==1.2.7
openai-whisper==20250625
requests==2.32.5
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
Flask==3.1.2
numpy==2.3.5
PyAudio==0.2.14
onnxruntime==1.24.1
torch==2.10.0
piper-tts==1.4.1
pathvalidate==3.3.1
openwakeword==0.6.0
pyttsx3==2.99
sounddevice==0.5.5
